+++ Sources/MapboxSearch/PublicAPI/Common/Models/Country/Country+ISO3166-1.swift
@@ -2,7 +2,7 @@
 // swiftlint:disable type_name
 
 extension Country {
-    enum ISO3166_1_alpha2: String {
+    public enum ISO3166_1_alpha2: String {
         case af = "AF"
         case ax = "AX"
         case al = "AL"